reorganize and refactor site JS to make it a little more scalable. use module pattern.
