+++ js/app/main.js
@@ -1,0 +1,5 @@
+var Map = require('./modules/map');
+
+jQuery( document ).ready(function( $ ) {
+  Map();
+});
